fix imports

+++ app/src/main/java/com/example/hashcache/models/Coordinate.java
@@ -4,7 +4,7 @@ package com.example.hashcache.models;
  * Represents a coordinate point with a latitude, longitude, and altitude
  */
 public class Coordinate {
-    private int[] coordinates = new int[3];
+    private double[] coordinates = new double[3];
 
     public Coordinate(int x, int y, int z){
         this.coordinates[0] = x;
@@ -12,7 +12,14 @@ public class Coordinate {
         this.coordinates[2] = z;
     }
 
-    protected void setCoordinate(int value, int position){
+    /**
+     * Sets the coordinate's latitude, longitude, or height altitude
+     * @param value The new coordinate value
+     * @param position The position of the value, 0 is longitude, 1 is latitude, and 2
+     *                 is altitude
+     * @throws IllegalArgumentException if the position is not 0, 1, or 2
+     */
+    protected void setCoordinate(double value, int position){
         if(position >= 0 && position < 3){
             this.coordinates[position] = value;
         }else{
@@ -21,7 +28,11 @@ public class Coordinate {
 
     }
 
-    public int[] getCoordinate(){
+    /**
+     * Get the coordinate values of this point
+     * @return this.coordinates The three coordinate points represented by this object
+     */
+    public double[] getCoordinate(){
         return this.coordinates;
     }
 }
